refactor(api): move quest security pre-handlers from shared
