--- api/src/quest/application/usecases/check-participation-belongs-to-combined-course.js
@@ -1,8 +1,0 @@
-import * as combinedCourseParticipationRepository from '../../infrastructure/repositories/combined-course-participation-repository.js';
-
-export const execute = async function ({ combinedCourseId, participationId }) {
-  return await combinedCourseParticipationRepository.isParticipationOnCombinedCourse({
-    combinedCourseId,
-    participationId,
-  });
-};
