--- api/src/quest/application/usecases/check-user-can-manage-combined-course.js
@@ -1,19 +1,0 @@
-import * as membershipRepository from '../../../shared/infrastructure/repositories/membership-repository.js';
-import * as combinedCourseRepository from '../../infrastructure/repositories/combined-course-repository.js';
-
-const execute = async function ({
-  userId,
-  combinedCourseId,
-  dependencies = { membershipRepository, combinedCourseRepository },
-}) {
-  const { organizationId } = await dependencies.combinedCourseRepository.getById({ id: combinedCourseId });
-
-  const memberships = await dependencies.membershipRepository.findByUserIdAndOrganizationId({
-    userId,
-    organizationId,
-  });
-
-  return memberships.length > 0;
-};
-
-export { execute };
