+++ api/src/quest/domain/usecases/can-manage-combined-course.js
@@ -1,0 +1,15 @@
+export async function canManageCombinedCourse({
+  userId,
+  combinedCourseId,
+  membershipRepository,
+  combinedCourseRepository,
+}) {
+  const { organizationId } = await combinedCourseRepository.getById({ id: combinedCourseId });
+
+  const memberships = await membershipRepository.findByUserIdAndOrganizationId({
+    userId,
+    organizationId,
+  });
+
+  return memberships.length > 0;
+}
