+++ api/src/quest/domain/usecases/has-access-to-combined-course.js
@@ -1,19 +1,20 @@
-import * as organizationApi from '../../../../src/organizational-entities/application/api/organization-api.js';
+import * as organizationApi from '../../../organizational-entities/application/api/organization-api.js';
 import { ORGANIZATION_FEATURE } from '../../../shared/domain/constants.js';
-import * as organizationFeatureRepository from '../../../shared/infrastructure/repositories/organization-feature-repository.js';
-import * as combinedCourseParticipantRepository from '../../infrastructure/repositories/combined-course-participant-repository.js';
-import * as combinedCourseRepository from '../../infrastructure/repositories/combined-course-repository.js';
 
-const execute = async function ({ userId, code }) {
+export async function hasAccessToCombinedCourse({
+  userId,
+  code,
+  combinedCourseRepository,
+  organizationFeatureRepository,
+  combinedCourseParticipantRepository,
+}) {
   const { organizationId } = await combinedCourseRepository.getByCode({ code });
-
   const hasImportFeature = await organizationFeatureRepository.isFeatureEnabledForOrganization({
     organizationId,
     featureKey: ORGANIZATION_FEATURE.LEARNER_IMPORT.key,
   });
 
   const organization = await organizationApi.getOrganization(organizationId);
-
   if (!hasImportFeature && !organization.isManagingStudents) {
     return true;
   }
@@ -23,6 +24,4 @@ const execute = async function ({ userId, code }) {
     organizationId,
   });
   return Boolean(organizationLearner);
-};
-
-export { execute };
+}
