+++ api/src/quest/domain/usecases/is-participation-on-combined-course.js
@@ -1,0 +1,10 @@
+export async function isParticipationOnCombinedCourse({
+  combinedCourseId,
+  participationId,
+  combinedCourseParticipationRepository,
+}) {
+  return await combinedCourseParticipationRepository.isParticipationOnCombinedCourse({
+    combinedCourseId,
+    participationId,
+  });
+}
