+++ api/src/shared/application/security-pre-handlers.js
@@ -7,11 +7,8 @@ import * as certificationIssueReportRepository from '../../certification/shared/
 import { Organization } from '../../organizational-entities/domain/models/Organization.js';
 import * as checkCampaignBelongsToCombinedCourseUsecase from '../../prescription/campaign/application/usecases/checkCampaignBelongsToCombinedCourse.js';
 import * as checkCampaignParticipationBelongsToUserUsecase from '../../prescription/campaign/application/usecases/checkCampaignParticipationBelongsToUser.js';
-import * as checkAuthorizationToAccessCombinedCourseUsecase from '../../quest/application/usecases/check-authorization-to-access-combined-course.js';
-import * as checkParticipationBelongsToCombinedCourseUsecase from '../../quest/application/usecases/check-participation-belongs-to-combined-course.js';
-import * as checkUserCanManageCombinedCourseUsecase from '../../quest/application/usecases/check-user-can-manage-combined-course.js';
 import { PIX_ADMIN } from '../../shared/domain/constants.js';
-import { ForbiddenAccess, NotFoundError } from '../domain/errors.js';
+import { ForbiddenAccess } from '../domain/errors.js';
 import { featureToggles } from '../infrastructure/feature-toggles/index.js';
 import * as organizationRepository from '../infrastructure/repositories/organization-repository.js';
 import { PromiseUtils } from '../infrastructure/utils/promise-utils.js';
@@ -636,72 +633,6 @@ async function checkAuthorizationToAccessCampaign(
   return _replyForbiddenError(h);
 }
 
-async function checkAuthorizationToAccessCombinedCourse(
-  request,
-  h,
-  dependencies = { checkAuthorizationToAccessCombinedCourseUsecase },
-) {
-  const userId = request.auth.credentials.userId;
-  const code = request.query?.filter?.code || request.params.code;
-
-  const canAccessCombinedCourse = await dependencies.checkAuthorizationToAccessCombinedCourseUsecase.execute({
-    userId,
-    code,
-  });
-
-  if (canAccessCombinedCourse) return h.response(true);
-  return _replyForbiddenError(h);
-}
-
-async function checkUserCanManageCombinedCourse(
-  request,
-  h,
-  dependencies = { checkUserCanManageCombinedCourseUsecase },
-) {
-  const userId = request.auth.credentials.userId;
-  const combinedCourseId = request.params.combinedCourseId;
-
-  try {
-    const canManageCombinedCourse = await dependencies.checkUserCanManageCombinedCourseUsecase.execute({
-      userId,
-      combinedCourseId,
-    });
-    if (canManageCombinedCourse) {
-      return h.response(true);
-    }
-    return _replyForbiddenError(h);
-  } catch (error) {
-    if (error instanceof NotFoundError) {
-      return _replyForbiddenError(h);
-    }
-    throw error;
-  }
-}
-
-async function checkParticipationBelongsToCombinedCourse(
-  request,
-  h,
-  dependencies = { checkParticipationBelongsToCombinedCourseUsecase },
-) {
-  const { participationId, combinedCourseId } = request.params;
-  try {
-    const isParticipationRelatedToCombinedCourse =
-      await dependencies.checkParticipationBelongsToCombinedCourseUsecase.execute({
-        participationId,
-        combinedCourseId,
-      });
-    if (isParticipationRelatedToCombinedCourse) {
-      return h.response(true);
-    }
-    return _replyForbiddenError(h);
-  } catch (error) {
-    if (error instanceof NotFoundError) {
-      return _replyForbiddenError(h);
-    }
-    throw error;
-  }
-}
-
 async function checkCombinedCoursesFeatureIsEnabled(request, h) {
   const areCombinedCoursesEnabled = await featureToggles.get('areCombinedCoursesEnabled');
   if (!areCombinedCoursesEnabled) {
@@ -921,8 +852,6 @@ export const securityPreHandlers = {
   checkAdminMemberHasRoleSupport,
   checkAuthorizationToManageCampaign,
   checkAuthorizationToAccessCampaign,
-  checkAuthorizationToAccessCombinedCourse,
-  checkUserCanManageCombinedCourse,
   checkCampaignBelongsToCombinedCourse,
   checkCertificationCenterIsNotScoManagingStudents,
   checkOrganizationHasFeature,
@@ -952,6 +881,5 @@ export const securityPreHandlers = {
   checkUserOwnsCertificationCourse,
   makeCheckOrganizationHasFeature,
   checkOrganizationAccess,
-  checkParticipationBelongsToCombinedCourse,
   checkCombinedCoursesFeatureIsEnabled,
 };
