+++ api/src/shared/dependencies.json
@@ -15,7 +15,6 @@
     "prescription/organization-place",
     "prescription/shared",
     "prescription/stages",
-    "quest",
     "school",
     "team"
   ],
