+++ api/tests/shared/acceptance/application/security-pre-handlers_test.js
@@ -1,9 +1,5 @@
 import { createServer } from '../../../../server.js';
 import { CombinedCourseBlueprint } from '../../../../src/quest/domain/models/CombinedCourseBlueprint.js';
-import {
-  OrganizationLearnerParticipationStatuses,
-  OrganizationLearnerParticipationTypes,
-} from '../../../../src/quest/domain/models/OrganizationLearnerParticipation.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
 import { ORGANIZATION_FEATURE } from '../../../../src/shared/domain/constants.js';
 import { Membership } from '../../../../src/shared/domain/models/Membership.js';
@@ -595,132 +591,6 @@ describe('Acceptance | Application | SecurityPreHandlers', function () {
         headers: generateAuthenticatedUserRequestHeaders({ userId }),
         method: 'GET',
         url: `/api/organizations/${organizationId}/place-statistics`,
-      };
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      expect(response.result).to.deep.equal(jsonApiError403);
-    });
-  });
-
-  describe('#checkAuthorizationToAccessCombinedCourse', function () {
-    let userId;
-    let organizationId;
-    let options;
-
-    beforeEach(async function () {
-      const code = 'COMBINIX1';
-      userId = databaseBuilder.factory.buildUser().id;
-      organizationId = databaseBuilder.factory.buildOrganization().id;
-      databaseBuilder.factory.buildCombinedCourse({ code, organizationId });
-
-      server.route({
-        method: 'GET',
-        path: '/api/test-combined-course',
-        handler: (r, h) => h.response({}).code(200),
-        config: {
-          pre: [
-            {
-              method: securityPreHandlers.checkAuthorizationToAccessCombinedCourse,
-            },
-          ],
-        },
-      });
-
-      options = {
-        headers: generateAuthenticatedUserRequestHeaders({ userId }),
-        method: 'GET',
-        url: `/api/test-combined-course?filter[code]=${code}`,
-      };
-
-      await databaseBuilder.commit();
-    });
-
-    it('should return true if connected user is allowed to access given combined course', async function () {
-      // given
-      databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return an error if connected user is not allowed to access given combined course', async function () {
-      // given
-      const importFeature = databaseBuilder.factory.buildFeature(ORGANIZATION_FEATURE.LEARNER_IMPORT);
-      databaseBuilder.factory.buildOrganizationFeature({
-        organizationId,
-        featureId: importFeature.id,
-      });
-
-      await databaseBuilder.commit();
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      expect(response.result).to.deep.equal(jsonApiError403);
-    });
-  });
-
-  describe('#checkParticipationBelongsToCombinedCourse', function () {
-    beforeEach(async function () {
-      server.route({
-        method: 'GET',
-        path: '/api/fake-cc/{combinedCourseId}/fake-p/{participationId}',
-        handler: (_, h) => h.response({}).code(200),
-        config: {
-          auth: false,
-          pre: [
-            {
-              method: securityPreHandlers.checkParticipationBelongsToCombinedCourse,
-            },
-          ],
-        },
-      });
-    });
-
-    it('should return true if participation belongs to combined course', async function () {
-      // given
-      const { id: combinedCourseId } = databaseBuilder.factory.buildCombinedCourse();
-      const { id: participationId } = databaseBuilder.factory.buildOrganizationLearnerParticipation({
-        combinedCourseId,
-        type: OrganizationLearnerParticipationTypes.COMBINED_COURSE,
-        status: OrganizationLearnerParticipationStatuses.STARTED,
-      });
-      await databaseBuilder.commit();
-
-      const options = {
-        method: 'GET',
-        url: `/api/fake-cc/${combinedCourseId}/fake-p/${participationId}`,
-      };
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return an error if participation does not belong to combined course', async function () {
-      // given
-      const { id: combinedCourseId } = databaseBuilder.factory.buildCombinedCourse();
-      const { id: participationId } = databaseBuilder.factory.buildOrganizationLearnerParticipation({
-        type: OrganizationLearnerParticipationTypes.COMBINED_COURSE,
-        status: OrganizationLearnerParticipationStatuses.STARTED,
-      });
-      await databaseBuilder.commit();
-
-      const options = {
-        method: 'GET',
-        url: `/api/fake-cc/${combinedCourseId}/fake-p/${participationId}`,
       };
 
       // when
