+++ api/tests/shared/unit/application/security-pre-handlers_test.js
@@ -2,7 +2,6 @@ import sinon from 'sinon';
 
 import { CampaignBelongsToCombinedCourseError } from '../../../../src/prescription/campaign/domain/errors.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
 import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
 import { expect } from '../../../test-helper.js';
 import { domainBuilder } from '../../../tooling/domain-builder/domain-builder.js';
@@ -2477,94 +2476,6 @@ describe('Shared | Unit | Application | SecurityPreHandlers', function () {
           5678,
           9999,
         );
-      });
-    });
-  });
-
-  describe('#checkAuthorizationToAccessCombinedCourseFromQuestId', function () {
-    let request;
-
-    beforeEach(function () {
-      request = {
-        auth: {
-          credentials: {
-            userId: 1234,
-          },
-        },
-        params: {
-          combinedCourseId: 'questId123',
-        },
-      };
-    });
-
-    context('Successful case', function () {
-      it('should authorize access when user is authorized to access combined course', async function () {
-        // given
-        const checkUserCanManageCombinedCourseUsecaseStub = {
-          execute: sinon.stub().resolves(true),
-        };
-
-        // when
-        const response = await securityPreHandlers.checkUserCanManageCombinedCourse(request, hFake, {
-          checkUserCanManageCombinedCourseUsecase: checkUserCanManageCombinedCourseUsecaseStub,
-        });
-
-        // then
-        expect(response.source).to.be.true;
-        expect(checkUserCanManageCombinedCourseUsecaseStub.execute).to.have.been.calledOnceWithExactly({
-          userId: 1234,
-          combinedCourseId: 'questId123',
-        });
-      });
-    });
-
-    context('Error cases', function () {
-      it('should forbid access when user is not authorized', async function () {
-        // given
-        const checkUserCanManageCombinedCourseUsecaseStub = {
-          execute: sinon.stub().resolves(false),
-        };
-
-        // when
-        const response = await securityPreHandlers.checkUserCanManageCombinedCourse(request, hFake, {
-          checkUserCanManageCombinedCourseUsecase: checkUserCanManageCombinedCourseUsecaseStub,
-        });
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid access when user is not authorized resource does not exist', async function () {
-        // given
-        const checkUserCanManageCombinedCourseUsecaseStub = {
-          execute: sinon.stub().rejects(new NotFoundError()),
-        };
-
-        // when
-        const response = await securityPreHandlers.checkUserCanManageCombinedCourse(request, hFake, {
-          checkUserCanManageCombinedCourseUsecase: checkUserCanManageCombinedCourseUsecaseStub,
-        });
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should throw if usecase return an unknown error', async function () {
-        // given
-        const WhatIsThisError = class extends Error {};
-        const checkUserCanManageCombinedCourseUsecaseStub = {
-          execute: sinon.stub().rejects(new WhatIsThisError()),
-        };
-
-        // when
-        const error = await catchErr(securityPreHandlers.checkUserCanManageCombinedCourse)(request, hFake, {
-          checkUserCanManageCombinedCourseUsecase: checkUserCanManageCombinedCourseUsecaseStub,
-        });
-
-        // then
-        expect(error).to.be.instanceOf(WhatIsThisError);
       });
     });
   });
